development script for server
