+++ package.json
@@ -2,10 +2,9 @@
   "name": "moneymanageapp",
   "version": "1.0.0",
   "description": "",
-  "main": "null",
   "scripts": {
     "start:client": "npm run start --prefix client",
-    "start:server": "npm run start --prefix server"
+    "start:server": "npm run dev --prefix server"
   },
   "repository": {
     "type": "git",
