+++ server/nodemon.json
@@ -1,0 +1,7 @@
+{
+    "verbose": true,
+    "ignore": ["*.test.js"],
+    "watch": ["src"],
+    "ext": "js",
+    "delay": "2"
+}
